backend/requirements.txt
python-dotenv
Django
Pillow
PyPDF2
pdf2image
reportlab
PyMuPDF
python-docx
httpx
pytesseract
scikit-learn
django-cors-headers

